Extract accepted activity helpers
